requirements.txt
flask
faiss-cpu
google-generativeai
numpy
python-dotenv